Enum pomelo_parse::syntax::SyntaxKind

#[repr(u16)]
pub enum SyntaxKind {

    EOF,
    UNKNOWN,
    ERROR,
    INFIX_OR_APP_EXP,
    INFIX_OR_APP_PAT,
    WHITESPACE,
    COMMENT,
    ABSTYPE_KW,
    AND_KW,
    ANDALSO_KW,
    AS_KW,
    CASE_KW,
    DATATYPE_KW,
    DO_KW,
    ELSE_KW,
    END_KW,
    EXCEPTION_KW,
    FN_KW,
    FUN_KW,
    HANDLE_KW,
    IF_KW,
    IN_KW,
    INFIX_KW,
    INFIXR_KW,
    LET_KW,
    LOCAL_KW,
    NONFIX_KW,
    OF_KW,
    OP_KW,
    OPEN_KW,
    ORELSE_KW,
    RAISE_KW,
    REC_KW,
    THEN_KW,
    TYPE_KW,
    VAL_KW,
    WITH_KW,
    WITHTYPE_KW,
    WHILE_KW,
    L_PAREN,
    R_PAREN,
    L_BRACKET,
    R_BRACKET,
    L_BRACE,
    R_BRACE,
    COMMA,
    COLON,
    SEMICOLON,
    ELLIPSIS,
    UNDERSCORE,
    PIPE,
    EQ,
    THICK_ARROW,
    THIN_ARROW,
    HASH,
    DOT,
    STAR,
    SCON_EXP,
    VID_EXP,
    RECORD_EXP,
    RECORD_SEL_EXP,
    LET_EXP,
    PAREN_EXP,
    EXP_ROW,
    TY_EXP,
    HANDLE_EXP,
    RAISE_EXP,
    FN_EXP,
    MATCH,
    MRULE,
    VAL_DEC,
    TY_DEC,
    DATATYPE_DEC,
    DATATYPE_REP,
    ABSTYPE_DEC,
    EXCEPT_DEC,
    LOCAL_DEC,
    OPEN_DEC,
    SEQ_DEC,
    INFIX_DEC,
    INFIXR_DEC,
    NONFIX_DEC,
    FIXITY,
    VAL_BIND,
    TY_BIND,
    DATA_BIND,
    CON_BIND,
    EX_BIND,
    EX_BIND_NOEQ,
    EX_BIND_EQ,
    WILDCARD_PAT,
    SCON_PAT,
    VID_PAT,
    RECORD_PAT,
    PAT_ROW,
    PAT_ROW_PAT,
    TY_PAT,
    LAYERED_PAT,
    TY,
    TYVAR_TY,
    RECORD_TY,
    CON_TY,
    FUN_TY,
    TY_ROW,
    DERIV_EXP,
    UNIT_EXP,
    TUPLE_EXP,
    ACCESSOR_EXP,
    CASE_MATCH_EXP,
    IF_EXP,
    ORELSE_EXP,
    ANDALSO_EXP,
    SEQ_EXP,
    WHILE_EXP,
    LIST_EXP,
    DERIV_PAT,
    UNIT_PAT,
    TUPLE_PAT,
    LIST_PAT,
    LAB_AS_VAR_PAT,
    LAB_AS_VAR_TY,
    LAB_AS_VAR_AS_PAT,
    TUPLE_TY_EXP,
    FVAL_BIND,
    FVAL_BIND_ROW,
    FUN_DEC,
    INT,
    REAL,
    WORD,
    CHAR,
    STRING,
    IDENT,
    VID,
    LONG_VID,
    TYVAR,
    TY_CON,
    LONG_TY_CON,
    LAB,
    STRID,
    LONG_STRID,
    FILE,

}

Kinds of syntax nodes in the Core SML language.

Variants

EOF

UNKNOWN

ERROR

INFIX_OR_APP_EXP

INFIX_OR_APP_PAT

WHITESPACE

COMMENT

ABSTYPE_KW

AND_KW

ANDALSO_KW

AS_KW

CASE_KW

DATATYPE_KW

DO_KW

ELSE_KW

END_KW

EXCEPTION_KW

FN_KW

FUN_KW

HANDLE_KW

IF_KW

IN_KW

INFIX_KW

INFIXR_KW

LET_KW

LOCAL_KW

NONFIX_KW

OF_KW

OP_KW

OPEN_KW

ORELSE_KW

RAISE_KW

REC_KW

THEN_KW

TYPE_KW

VAL_KW

WITH_KW

WITHTYPE_KW

WHILE_KW

L_PAREN

R_PAREN

L_BRACKET

R_BRACKET

L_BRACE

R_BRACE

COMMA

COLON

SEMICOLON

ELLIPSIS

UNDERSCORE

PIPE

EQ

THICK_ARROW

THIN_ARROW

HASH

DOT

STAR

SCON_EXP

VID_EXP

RECORD_EXP

RECORD_SEL_EXP

LET_EXP

PAREN_EXP

EXP_ROW

TY_EXP

HANDLE_EXP

RAISE_EXP

FN_EXP

MATCH

MRULE

VAL_DEC

TY_DEC

DATATYPE_DEC

DATATYPE_REP

ABSTYPE_DEC

EXCEPT_DEC

LOCAL_DEC

OPEN_DEC

SEQ_DEC

INFIX_DEC

INFIXR_DEC

NONFIX_DEC

FIXITY

VAL_BIND

TY_BIND

DATA_BIND

CON_BIND

EX_BIND

EX_BIND_NOEQ

EX_BIND_EQ

WILDCARD_PAT

SCON_PAT

VID_PAT

RECORD_PAT

PAT_ROW

PAT_ROW_PAT

TY_PAT

LAYERED_PAT

TY

TYVAR_TY

RECORD_TY

CON_TY

FUN_TY

TY_ROW

DERIV_EXP

UNIT_EXP

TUPLE_EXP

ACCESSOR_EXP

CASE_MATCH_EXP

IF_EXP

ORELSE_EXP

ANDALSO_EXP

SEQ_EXP

WHILE_EXP

LIST_EXP

DERIV_PAT

UNIT_PAT

TUPLE_PAT

LIST_PAT

LAB_AS_VAR_PAT

LAB_AS_VAR_TY

LAB_AS_VAR_AS_PAT

TUPLE_TY_EXP

FVAL_BIND

FVAL_BIND_ROW

FUN_DEC

INT

REAL

WORD

CHAR

STRING

IDENT

VID

LONG_VID

TYVAR

TY_CON

LONG_TY_CON

LAB

STRID

LONG_STRID

FILE

Implementations

impl SyntaxKind

pub fn is_special_constant(&self) -> bool

Returns true if self is a special constant (literal).

pub fn is_trivia(&self) -> bool

Returns true if self is trivia (whitespace or comment).

pub fn is_dec_kw(&self) -> bool

Returns true if self is a keyword signalling a declaration.

pub fn is_ty_atom(&self) -> bool

Returns true if self is the start of a atomic type (either a type variable or the start of a record or tuple type).

pub fn is_kw(&self) -> bool

Returns true if self is a keyword.

pub fn is_atomic_exp_start(&self) -> bool

Returns true if self is the start of an atomic expression.

pub fn is_atomic_pat_start(&self) -> bool

Returns true if self is the start of an atomic pattern.

pub fn from_keyword(s: &str) -> Option<Self>

Returns Some(kind) if s is a keyword, else None.

pub fn from_symbol(s: &str) -> Option<Self>

Returns Some(kind) if s is a symbol, else None.

pub fn from_lexed(lexkind: LexKind, text: &str) -> (Self, Option<&'static str>)

Converts a lex token to a parse token.

Note that the return type is a (val, err) tuple instead of a Result. This is because we want to collect errors here but defer their handling until after everything has been parsed.